First commit and setup of Github
This was a very thorny setup!

We had many permissions errors when trying to save files in the Atom
text editor and then when trying to make a `git commit`.

For future reference, when making a new project, follow the following directions:

1. Open Terminal.
2. type `cd projects` and hit enter
3. type `mkdir mynewprojectdirectoryname` and hit enter
4. type ls -as and hit enter
5. You should see the mynewprojectdirectoryname in the list of files that is shown when you type `ls -as`. If you do not see `mynewprojectdirectoryname`type `cd ..` and then type `ls -as` again.
6. When you see `mynewprojectdirectoryname` type `sudo su` at the command line. If you are asked for a password, login with the password for your laptop.
7. type `chmod 777 mynewprojectdirectoryname` and hit enter
8. Type ls -a and be sure that all permissions read rwx (read, write, execute) and then type `cd mynewprojectdirectoryname` and then ls -a to be sure any files in your directory also show rwx.

VERY IMPORTANT! Once you finish changing permissions, you must log out of root. Type `exit` and then `whoami` to be sure it says your username rather than `root`.  If you ever get the error `cannot save permission denied`, go back to the terminal and change the permissions on the folder again.

Thank You Gladys